finished delete review

+++ backend/db/migrations/20231127204226-create-spot.js
@@ -55,7 +55,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       price: {
-        type: Sequelize.DOUBLE,
+        type: Sequelize.DOUBLE(10, 2),
         allowNull: false
       },
       createdAt: {
